Move `EngineActor` to its own file & create wrapper struct
